favorite_page

+++ resources/views/microposts/microposts.blade.php
@@ -17,16 +17,18 @@
                     <div class="container">
                         <div class="row">
                         {{-- お気に入りボタン --}}
-                        <div class="col-md-1.5">@include('favorites.favorite_button')</div>
-                        <div class="col-md-1.5">@if (Auth::id() == $micropost->user_id)
+                        <div class="col-md-1.5">
+                            @include('favorites.favorite_button')</div>
+                        @if (Auth::id() == $micropost->user_id)
                             {{-- 投稿削除ボタンのフォーム --}}
+                            <div class="col-md-1.5">
                             {!! Form::open(['route' => ['microposts.destroy', $micropost->id], 'method' => 'delete']) !!}
                                 {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
                             {!! Form::close() !!}
-                        @endif</div>
+                        @endif
                         </div>
-                        <!-- cssの呼び出し -->
-                        <link rel="stylesheet" href="{{ asset('css/style.css') }}">
+                        </div>
+                        
                     </div>
                 </div>
             </li>
